requirements/local.txt
# local development dependencies go here
-r base.txt

django-debug-toolbar
django-autofixture
django-webtest